feat(BUX-166): Retrieve Merkle Path from Arc
