+++ broadcast/internal/arc/mocks/consts.go
@@ -47,31 +47,45 @@ var (
 	}
 
 	SubmittedTx = &broadcast_api.SubmittedTx{
-		Status:      fixtures.TxResponseStatus,
-		Title:       fixtures.TxResponseTitle,
-		TxStatus:    fixtures.TxStatus,
-		BlockHash:   fixtures.TxBlockHash,
-		BlockHeight: fixtures.TxBlockHeight,
-		ExtraInfo:   fixtures.TxExtraInfo,
+		Status: fixtures.TxResponseStatus,
+		Title:  fixtures.TxResponseTitle,
+		BaseTxResponse: broadcast_api.BaseTxResponse{
+			BlockHash:   fixtures.TxBlockHash,
+			BlockHeight: fixtures.TxBlockHeight,
+			ExtraInfo:   fixtures.TxExtraInfo,
+			MerklePath:  fixtures.TxMerklePath,
+			Timestamp:   fixtures.Timestamp,
+			TxStatus:    fixtures.TxStatus,
+			TxID:        fixtures.TxId,
+		},
 	}
 
 	SubmittedTxSecondary = &broadcast_api.SubmittedTx{
-		Status:      fixtures.TxResponseStatus,
-		Title:       fixtures.TxResponseTitle,
-		TxStatus:    fixtures.TxStatus,
-		BlockHash:   fixtures.TxBlockHashSecondary,
-		BlockHeight: fixtures.TxBlockHeightSecondary,
-		ExtraInfo:   fixtures.TxExtraInfo,
+		Status: fixtures.TxResponseStatus,
+		Title:  fixtures.TxResponseTitle,
+		BaseTxResponse: broadcast_api.BaseTxResponse{
+			BlockHash:   fixtures.TxBlockHashSecondary,
+			BlockHeight: fixtures.TxBlockHeightSecondary,
+			ExtraInfo:   fixtures.TxExtraInfo,
+			MerklePath:  fixtures.TxMerklePath,
+			Timestamp:   fixtures.Timestamp,
+			TxStatus:    fixtures.TxStatus,
+			TxID:        fixtures.TxIdSecondary,
+		},
 	}
 )
 
 func QueryTx(txID string) *broadcast_api.QueryTxResponse {
 	return &broadcast_api.QueryTxResponse{
 		BaseResponse: broadcast_api.BaseResponse{Miner: fixtures.ProviderMain},
-		Timestamp:    fixtures.Timestamp,
-		TxStatus:     fixtures.TxStatus,
-		BlockHash:    fixtures.TxBlockHash,
-		BlockHeight:  fixtures.TxBlockHeight,
-		TxID:         txID,
+		BaseTxResponse: broadcast_api.BaseTxResponse{
+			BlockHash:   fixtures.TxBlockHash,
+			BlockHeight: fixtures.TxBlockHeight,
+			ExtraInfo:   fixtures.TxExtraInfo,
+			MerklePath:  fixtures.TxMerklePath,
+			Timestamp:   fixtures.Timestamp,
+			TxStatus:    fixtures.TxStatus,
+			TxID:        txID,
+		},
 	}
 }
