+++ examples/submit_transaction/submit_tx.go
@@ -12,7 +12,7 @@ import (
 func main() {
 	token := ""
 	apiURL := "https://tapi.taal.com/arc"
-	tx := broadcast.Transaction{Hex: "0100000001d6d1607b208b30c0a3fe21d563569c4d2a0f913604b4c5054fe267da6be324ab220000006b4830450221009a965dcd5d42983090a63cfd761038ff8adcea621c46a68a205f326292a95383022061b8d858f366c69f3ebd30a60ccafe36faca4e242ac3d2edd3bf63b669bcf23b4121034e871e147aa4a3e2f1665eaf76cf9264d089b6a91702af92bd6ce33bac84a765ffffffff0123020000000000001976a914d8819a7197d3e221e15f4348203fdecfd29fa2b888ac00000000"}
+	tx := broadcast.Transaction{Hex: "0100000001b71cbd19dd20160d3e662b2f10f686dba7ffad05caa687f0a9950a8efb394fbf060000006a473044022005dfb6c03648ce8cc6dbf3e44f905ff26f4140fec1868a237685a31e8df46c3e02204a0b9da3fb3514946cda2f2eab4662217633de10d4a83865c376ba39e093e373412103b671ba3ef17b2b1eea224579af868d40ec6645d819461b82eaf7846ae281d455ffffffff08f4010000000000001976a91449674dc635418b384086ef9a95100cc40e92799e88acf4010000000000001976a914bf062a0a6915954069141ad7584506997ad2bddf88acf4010000000000001976a914cb9a17dfa0c033cac9f297d9f82add28673e8f1b88ac88130000000000001976a914a58d30914be2cddffbdb2cfd544445729f993f0388acf4010000000000001976a914f6f6346b6f4a8c0b3dc1e3ffd7ba897018e7059188acf4010000000000001976a914640f71761a00fbebedeba33bc41f18bcfdd0118088ac0000000000000000fd2401006a22314d6a4a7251744a735959647a753254487872654e69524c53586548637a417778550a5361746f73656e6465720a746578742f706c61696e057574662d38017c223150755161374b36324d694b43747373534c4b79316b683536575755374d7455523503534554036170700a7361746f73656e64657204747970650c7265676973746572557365720c7061796d61696c204861736840373863613031353837646635373533323265623538386233623830623433313239633763636435383266343361303662623538363133353233323737613663340970686f6e65486173684039643535626465323661613139343364326234366463303665343735356139363536666564323062626632643839313138353339346139623431633034643539c3090000000000001976a914397bafc4137c7d04f77afcb42e80f6955dde3a4188ac00000000"}
 
 	cfg := broadcast_client.ArcClientConfig{
 		Token:  token,
@@ -28,7 +28,5 @@ func main() {
 		log.Fatalf("error: %s", err.Error())
 	}
 
-	log.Printf("miner: %s", result.Miner)
-	log.Printf("hash: %s", result.BlockHash)
-	log.Printf("status: %s", result.TxStatus)
+	broadcast.PrettyPrint("Result: ", result)
 }
